#47 [refactor] : 날짜 함수 uitils로 분리
